+++ src/main.jsx
@@ -3,7 +3,6 @@ import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { StrictMode } from "react";
 import App from "./App";
 import SearchPages from "./pages/Search.jsx";
-import Finish from "./pages/Finish.jsx";
 import Chat from "./pages/Chat/index.jsx";
 import Backend from "./pages/Backend.jsx";
 import "./index.css";
@@ -15,7 +14,6 @@ createRoot(document.getElementById("root")).render(
         <Route path="/" element={<App />}>
           <Route index element={<SearchPages />} />
           <Route path="chat" element={<Chat />} />
-          <Route path="finishPage" element={<Finish />} />
           <Route path="backend" element={<Backend />} />
           <Route path="*" element={<Navigate to="/" replace />} />
         </Route>
