+++ src/pages/Chat/Checkout.jsx
@@ -4,10 +4,10 @@ import PropTypes from "prop-types";
 
 function Checkout({ checkout }) {
   const state = useContext(ChatContext);
-  const dispatch = useContext(ChatDispatchContext);
+  const { dispatch } = useContext(ChatDispatchContext);
 
   return (
-    <div className={`${state.isPurchase ? "flex" : "hidden"} justify-center items-center w-container h-full fixed top-0`}>
+    <div className={`${state.isPerchase ? "flex" : "hidden"} justify-center items-center w-container h-full fixed top-0`}>
       <div className="w-64 h-84 bg-white mx-auto py-2 px-4 flex flex-col gap-4 text-sm rounded-xl">
         <h4 className="text-center font-bold leading-normal text-lg text-primary-800 mt-2">訂單即將送出</h4>
         <p className="text-center font-bold">已確認品項、數量並進行結帳嗎？</p>
