+++ src/pages/Chat/index.jsx
@@ -1,22 +1,23 @@
 import { useEffect, useContext } from "react";
-import { ChatContext, ChatDispatchContext } from "../chatContext";
-import responses from "./responses.json";
-import { db, storage, collection, addDoc, query, orderBy, onSnapshot, serverTimestamp, doc, getDocs, where, ref, uploadBytesResumable, getDownloadURL } from "../utils/firebase";
-import useGoogleVisionAPI from "../utils/useGoogleVisionAPI";
-import tappay from "../utils/tappay";
+import { ChatContext, ChatDispatchContext } from "../../chatContext";
+
+import { db, storage, collection, addDoc, query, orderBy, onSnapshot, serverTimestamp, getDocs, ref, uploadBytesResumable, getDownloadURL } from "../../utils/firebase";
+import { fetchOrderInfo, fetchProductInfo, fetchGPT } from "../../utils/fetch";
+import useGoogleVisionAPI from "../../utils/useGoogleVisionAPI";
+import tappay from "../../utils/tappay";
 import { format, isToday, isYesterday, differenceInMinutes } from "date-fns";
 import { zhTW } from "date-fns/locale";
 import "react-photo-view/dist/react-photo-view.css";
 
-import Header from "./Chat/Header";
-import OrderCard from "./Chat/OrderCard";
-import ShopCard from "./Chat/ShopCard";
-import DateLabel from "./Chat/DateLabel";
-import ChatSection from "./Chat/ChatSection";
-import Choose from "./Chat/Choose";
-import Checkout from "./Chat/Checkout";
-import Order from "./Chat/Order";
-import TypeIn from "./Chat/TypeIn";
+import Header from "./Header";
+import OrderCard from "./OrderCard";
+import ShopCard from "./ShopCard";
+import DateLabel from "./DateLabel";
+import ChatSection from "./ChatSection";
+import Choose from "./Choose";
+import Checkout from "./Checkout";
+import Order from "./Order";
+import TypeIn from "./TypeIn";
 
 function Chat() {
   const { labels, handleAnalyzeImage } = useGoogleVisionAPI();
@@ -28,79 +29,17 @@ function Chat() {
   const state = useContext(ChatContext);
   const { dispatch, scrollToBottom } = useContext(ChatDispatchContext);
 
-  const fetchOrderInfo = async (shopId, orderNumber) => {
-    try {
-      const ordersCollectionRef = collection(db, "orders");
-      const orderQuery = query(ordersCollectionRef, where("shopId", "==", shopId), where("orderNumber", "==", orderNumber));
-      const querySnapshot = await getDocs(orderQuery);
-      if (!querySnapshot.empty) {
-        const orderDoc = querySnapshot.docs[0];
-        dispatch({ type: "SET_ORDER_INFO", payload: orderDoc.data() });
-      } else {
-        console.log("No matching order found!");
-      }
-    } catch (error) {
-      console.error("Error getting order documents:", error);
-    }
-  };
-
-  const fetchProductInfo = async (shopId, productNumber) => {
-    try {
-      const shopQuery = query(collection(db, "shops"), where("shopId", "==", shopId));
-      const shopSnapshot = await getDocs(shopQuery);
-      if (!shopSnapshot.empty) {
-        const shopDoc = shopSnapshot.docs[0];
-        const shopDocId = shopDoc.id;
-        const shopName = shopDoc.data().shopName;
-
-        dispatch({ type: "SET_SHOP_NAME", payload: shopName });
-
-        const productsCollectionRef = collection(doc(db, "shops", shopDocId), "products");
-
-        if (productNumber) {
-          const productQuery = query(productsCollectionRef, where("productNumber", "==", productNumber));
-
-          const productSnapshot = await getDocs(productQuery);
-          if (!productSnapshot.empty) {
-            const productDoc = productSnapshot.docs[0];
-            dispatch({ type: "SET_PRODUCT_INFO", payload: productDoc.data() });
-          } else {
-            console.log("No matching product found!");
-            dispatch({ type: "SET_PRODUCT_INFO", payload: null });
-          }
-        } else {
-          const productSnapshot = await getDocs(productsCollectionRef);
-
-          const productList = productSnapshot.docs.map((doc) => doc.data());
-
-          const randomIndex = Math.floor(Math.random() * productList.length);
-
-          const productDoc = productSnapshot.docs[randomIndex];
-          dispatch({ type: "SET_PRODUCT_INFO", payload: productDoc.data() });
-        }
-
-        dispatch({ type: "TOGGLE_SHOP_INFO", payload: true });
-      } else {
-        console.log("No matching shop found for the given shopId!");
-        dispatch({ type: "SET_SHOP_NAME", payload: "商家名稱未找到" });
-        dispatch({ type: "TOGGLE_SHOP_INFO", payload: true });
-      }
-    } catch (error) {
-      console.error("Error getting product documents:", error);
-    }
-  };
-
   useEffect(() => {
     if (shopId) {
       dispatch({ type: "TOGGLE_SHOP_INFO", payload: true });
-      fetchProductInfo(shopId, productNumber);
+      fetchProductInfo(shopId, productNumber, dispatch);
       if (orderNumber) {
         dispatch({ type: "TOGGLE_ORDER_INFO", payload: true });
-        fetchOrderInfo(shopId, orderNumber);
+        fetchOrderInfo(shopId, orderNumber, dispatch);
       } else {
         dispatch({ type: "TOGGLE_SHOP_INFO", payload: true });
         dispatch({ type: "TOGGLE_PRODUCT_INFO", payload: true });
-        fetchProductInfo(shopId, productNumber);
+        fetchProductInfo(shopId, productNumber, dispatch);
       }
     } else {
       dispatch({ type: "TOGGLE_ORDER_INFO", payload: false });
@@ -131,8 +70,8 @@ function Chat() {
       await tappay.setupSDK();
       tappay.setupCard();
     };
-
     setupTappay();
+
     handleScroll();
     window.addEventListener("scroll", handleScroll);
 
@@ -170,6 +109,7 @@ function Chat() {
   let scrollTimeout;
 
   useEffect(() => {
+    scrollToBottom();
     if (state.messages.length > 0) {
       const handleScroll = () => {
         dispatch({ type: "SET_IS_SCROLLING", payload: true });
@@ -219,8 +159,8 @@ function Chat() {
     }
   }, [state.messages]);
 
-  const predefinedResponses = responses.map((item) => ({
-    pattern: new RegExp(item.pattern, "i"),
+  const predefinedResponses = state.faqs.map((item) => ({
+    pattern: new RegExp(item.keyword, "i"),
     response: item.response,
   }));
 
@@ -231,50 +171,6 @@ function Chat() {
       from: from,
       isUsefull: "",
     });
-  };
-
-  const fetchGPT = async (inputText, document) => {
-    const apiKey = import.meta.env.VITE_OPENAI_API_KEY;
-    const apiUrl = "https://api.openai.com/v1/chat/completions";
-
-    try {
-      const res = await fetch(apiUrl, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${apiKey}`,
-        },
-        body: JSON.stringify({
-          model: "gpt-4o-mini",
-          max_tokens: 150,
-          messages: [
-            {
-              role: "system",
-              content: "你是一個全程使用繁體中文並且非常人性化回覆「已登入」的使用者提問MOMO電商客服相關問題，且不會提到「關鍵字」三個字的富邦媒體電商客服人員",
-            },
-            {
-              role: "user",
-              content: inputText,
-            },
-          ],
-          temperature: 0.7,
-        }),
-      });
-
-      if (res.ok) {
-        const data = await res.json();
-        await addMessage(document, data.choices[0].message.content, "shop");
-      } else if (res.status === 429) {
-        console.error("Too many requests. Please try again later.");
-        dispatch({ type: "SET_GPT_ERROR", payload: "Too many requests. Please try again later." });
-      } else {
-        console.error("Error:", res.status, res.statusText);
-        dispatch({ type: "SET_GPT_ERROR", payload: "An error occurred. Please try again later." });
-      }
-    } catch (error) {
-      console.error("Fetch error:", error);
-      dispatch({ type: "SET_GPT_ERROR", payload: "An error occurred. Please try again later." });
-    }
   };
 
   const sendMessage = async (url) => {
@@ -361,6 +257,20 @@ function Chat() {
       }, 1000);
     }
   }, [labels]);
+
+  useEffect(() => {
+    const fetchFaqs = async () => {
+      const querySnapshot = await getDocs(collection(db, "faq"));
+      const faqList = querySnapshot.docs.map((doc) => ({
+        keyword: doc.data().keyword,
+        response: doc.data().response,
+      }));
+
+      dispatch({ type: "SET_FAQS", payload: faqList });
+    };
+
+    fetchFaqs();
+  }, []);
 
   async function checkout() {
     try {
